Working pipes

+++ include/cnets.h
@@ -6,7 +6,7 @@
 #include <sys/socket.h>
 #include "ipv4h.h"
 
-uint16_t raw_checksum(const void* addr, size_t count, uint32_t);
+uint16_t raw_checksum(const void* addr, size_t count, uint64_t);
 
 uint16_t checksum(struct IpHeader* header);
 uint16_t tcp_udp_checksum(uint32_t saddr, uint32_t daddr, uint8_t proto,
